Moved shaders to a folder for organization purposes.
Updated README.txt to reflect the new organization for the shaders.

Added error querying to display, currently will only output an error
message should one be posted.

Changed phong.vs to hopefully handle the error posted by Professor Shafae,
now it explicitely gets the xyz components of a vertex's eye coordinates
as opposed to assuming the converstion from a vec4 to a vec3 is automatic.